solving some lint issues
