+++ .prettierrc
@@ -3,5 +3,6 @@
   "trailingComma": "all",
   "singleQuote": true,
   "printWidth": 80,
-  "useTabs": true
+  "useTabs": false,
+  "tabWidth": 2
 }
