+++ src/App.tsx
@@ -3,42 +3,44 @@ import React from 'react'
 import { LanguageContextProvider } from 'contexts/LanguageContext'
 import { NavigationContainer } from '@react-navigation/native'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
-import { EnterScreen } from 'screens/EnterScreen'
+import { EnterScreen } from 'screens/Enter'
 import { ThemeContextProvider } from 'contexts/ThemeContext'
 import { useCustomFonts } from 'hooks/useCustomFonts'
 import { Feed } from 'screens/Feed'
 import { themes } from './themes'
 
 export type RootStackParamList = {
-	Enter: undefined
-	Feed: undefined
+  Enter: undefined
+  Feed: undefined
+  NewPost: undefined
+  NewPet: undefined
 }
 
 const Stack = createNativeStackNavigator<RootStackParamList>()
 
 const App = () => {
-	const [fontsLoaded] = useCustomFonts()
-	if (!fontsLoaded) {
-		return null
-	}
+  const [fontsLoaded] = useCustomFonts()
+  if (!fontsLoaded) {
+    return null
+  }
 
-	return (
-		<LanguageContextProvider>
-			<ThemeContextProvider theme={themes.snapcat}>
-				<NavigationContainer>
-					<Stack.Navigator
-						initialRouteName="Enter"
-						screenOptions={{
-							headerShown: false,
-						}}
-					>
-						<Stack.Screen name="Enter" component={EnterScreen} />
-						<Stack.Screen name="Feed" component={Feed} />
-					</Stack.Navigator>
-				</NavigationContainer>
-			</ThemeContextProvider>
-		</LanguageContextProvider>
-	)
+  return (
+    <LanguageContextProvider>
+      <ThemeContextProvider theme={themes.snapcat}>
+        <NavigationContainer>
+          <Stack.Navigator
+            initialRouteName="Enter"
+            screenOptions={{
+              headerShown: false,
+            }}
+          >
+            <Stack.Screen name="Enter" component={EnterScreen} />
+            <Stack.Screen name="Feed" component={Feed} />
+          </Stack.Navigator>
+        </NavigationContainer>
+      </ThemeContextProvider>
+    </LanguageContextProvider>
+  )
 }
 
 export default registerRootComponent(App)
