+++ src/components/Button/types.ts
@@ -1,8 +1,8 @@
 export type ButtonProps = {
-	backgroundColor?: string
-	color?: string
-	hasShadow?: boolean
-	isBold?: boolean
-	text: string
-	onClick: () => void
+  backgroundColor?: string
+  color?: string
+  hasShadow?: boolean
+  isBold?: boolean
+  text: string
+  onClick: () => void
 }
