+++ src/contexts/LanguageContext/types.ts
@@ -1,6 +1,6 @@
 import { Language, Translation } from 'translations/types'
 
 export type LanguageContextContent = {
-	translation: Translation
-	setLanguage: (language: Language) => void
+  translation: Translation
+  setLanguage: (language: Language) => void
 }
