+++ src/contexts/ThemeContext/index.tsx
@@ -5,11 +5,11 @@ import { themes } from 'themes/index'
 const ThemeContext = createContext<Theme>(themes.snapcat)
 
 export const ThemeContextProvider = (props: ThemeContextProviderProps) => {
-	return (
-		<ThemeContext.Provider value={props.theme}>
-			{props.children}
-		</ThemeContext.Provider>
-	)
+  return (
+    <ThemeContext.Provider value={props.theme}>
+      {props.children}
+    </ThemeContext.Provider>
+  )
 }
 
 export const useTheme = () => React.useContext(ThemeContext)
