+++ src/hooks/useCustomFonts/index.ts
@@ -1,9 +1,9 @@
 import { useFonts } from 'expo-font'
 
 export const useCustomFonts = () =>
-	useFonts({
-		'Poppins-Regular': require('../../../assets/fonts/Poppins-Regular.ttf'),
-		'Poppins-Bold': require('../../../assets/fonts/Poppins-Bold.ttf'),
-		'Poppins-Medium': require('../../../assets/fonts/Poppins-Medium.ttf'),
-		'Poppins-SemiBold': require('../../../assets/fonts/Poppins-SemiBold.ttf'),
-	})
+  useFonts({
+    'Poppins-Regular': require('../../../assets/fonts/Poppins-Regular.ttf'),
+    'Poppins-Bold': require('../../../assets/fonts/Poppins-Bold.ttf'),
+    'Poppins-Medium': require('../../../assets/fonts/Poppins-Medium.ttf'),
+    'Poppins-SemiBold': require('../../../assets/fonts/Poppins-SemiBold.ttf'),
+  })
