+++ src/hooks/useCustomFonts/types.ts
@@ -1,6 +1,6 @@
 export enum Font {
-	Poppins_Regular = 'Poppins-Regular',
-	Poppins_Bold = 'Poppins-Bold',
-	Poppins_Medium = 'Poppins-Medium',
-	Poppins_SemiBold = 'Poppins-SemiBold',
+  Poppins_Regular = 'Poppins-Regular',
+  Poppins_Bold = 'Poppins-Bold',
+  Poppins_Medium = 'Poppins-Medium',
+  Poppins_SemiBold = 'Poppins-SemiBold',
 }
