+++ src/screens/Enter/types.ts
@@ -1,0 +1,7 @@
+import { NativeStackScreenProps } from '@react-navigation/native-stack'
+import { RootStackParamList } from 'src/App'
+
+export type EnterScreenProps = NativeStackScreenProps<
+  RootStackParamList,
+  'Enter'
+>
