+++ src/screens/NewPet/index.tsx
@@ -1,14 +1,14 @@
 import { SafeAreaView, Text } from 'react-native'
 import React from 'react'
-import { FeedProps } from './types'
 import { TitleBar } from 'components/TitleBar'
+import { NewPetScreenProps } from './types'
 
-export const Feed = ({ navigation }: FeedProps) => {
+export const NewPetScreen = ({ navigation }: NewPetScreenProps) => {
   navigation
   return (
     <SafeAreaView>
       <TitleBar />
-      <Text>Feed</Text>
+      <Text>New Pet</Text>
     </SafeAreaView>
   )
 }
