+++ src/screens/NewPet/types.ts
@@ -1,7 +1,7 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack'
 import { RootStackParamList } from 'src/App'
 
-export type EnterScreenProps = NativeStackScreenProps<
-	RootStackParamList,
-	'Enter'
+export type NewPetScreenProps = NativeStackScreenProps<
+  RootStackParamList,
+  'NewPet'
 >
