+++ src/themes/index.ts
@@ -1,5 +1,5 @@
 import { snapcatTheme } from './snapcat'
 
 export const themes = {
-	snapcat: snapcatTheme,
+  snapcat: snapcatTheme,
 }
