+++ src/themes/snapcat.ts
@@ -1,12 +1,12 @@
 import { Theme } from 'contexts/ThemeContext/types'
 
 export const snapcatTheme: Theme = {
-	primary: '#8A29E8',
-	secondary: '#FAE360',
-	terciary: '#8FBDDF',
-	black: '#000000',
-	white: '#FFFFFF',
-	gray: '#F0F1DF',
-	enterLogo: require('../../assets/SnapCat/enter.png'),
-	titleLogo: require('../../assets/SnapCat/title.png'),
+  primary: '#8A29E8',
+  secondary: '#FAE360',
+  terciary: '#8FBDDF',
+  black: '#000000',
+  white: '#FFFFFF',
+  gray: '#F0F1DF',
+  enterLogo: require('../../assets/SnapCat/enter.png'),
+  titleLogo: require('../../assets/SnapCat/title.png'),
 }
