+++ src/translations/languages/en.ts
@@ -1,5 +1,5 @@
 import { Translation } from 'translations/types'
 
 export const en: Translation = {
-	enter: 'enter',
+  enter: 'enter',
 }
