+++ src/translations/languages/pt.ts
@@ -1,5 +1,5 @@
 import { Translation } from 'translations/types'
 
 export const pt: Translation = {
-	enter: 'entrar',
+  enter: 'entrar',
 }
